web/requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
sqlparse==0.4.3
tzdata==2023.3

